Refactor: move entry methods to dspace-client.service
